+++ scripts/04-calculate-rs.py
